refactor configuration management
