+++ __tests__/integration/mirador/index.html
@@ -11,7 +11,15 @@
     <script>document.write("<script type='text/javascript' src='../../../dist/mirador.min.js?v=" + Date.now() + "'><\/script>");</script>
     <script type="text/javascript">
      var miradorInstance = Mirador.viewer({
-       id: 'mirador'
+       id: 'mirador',
+       windows: [{
+         id: 'window-1',
+         loadedManifest: 'https://iiif.harvardartmuseums.org/manifests/object/299843',
+       },
+       {
+         id: 'window-2',
+         loadedManifest: 'https://iiif.harvardartmuseums.org/manifests/object/299843'
+       }]
      });
     </script>
   </body>
