+++ src/config/settings.js
@@ -1,4 +1,4 @@
 export default {
-  foo: 'bar',
+  plugins: [],
   thumbnailNavigationHeight: 150,
 };
