+++ src/store.js
@@ -2,13 +2,11 @@
 // redux modules for nested stores
 // state normalisation
 // (normalizer library)
-
 import thunkMiddleware from 'redux-thunk';
 import { createStore, applyMiddleware } from 'redux';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import createRootReducer from './reducers/index';
 import * as ActionCreators from './actions';
-
 
 /**
  * Configure Store
